requirements.txt
altgraph
certifi
charset-normalizer
idna
lxml
packaging
pefile
pyinstaller
pyinstaller-hooks-contrib
PyMuPDF
PySide6
python-docx
pywin32-ctypes
requests
setuptools
shiboken6
typing_extensions
urllib3